feat: Refactor charge display components and update API methods
- Replace PriceValueCell with ChargeValueCell in multiple components for consistent charge rendering.
- Introduce ChargeValueCell component to handle tiered pricing display with tooltips for better user experience.
- Update UserApi to use PUT method for updating tenant details, ensuring correct API usage.
- Adjust billing model display logic in transformed_plan to enhance clarity in charge representation.
