+++ src/components/organisms/PlanForm/UsageChargePreview.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { Pencil, Trash2 } from 'lucide-react';
 import { toSentenceCase } from '@/utils/common/helper_functions';
 import { InternalPrice } from './SetupChargesSection';
-import ChargeValueCell from '@/pages/product-catalog/plans/PriceValueCell';
+import ChargeValueCell from '@/pages/product-catalog/plans/ChargeValueCell';
 
 interface Props {
 	charge: InternalPrice;
