+++ src/components/organisms/PlanForm/VolumeTieredPricingForm.tsx
@@ -1,9 +1,10 @@
 import { FC } from 'react';
 import { cn } from '@/lib/utils';
 import { RiDeleteBin6Line } from 'react-icons/ri';
-import { Button, Input, Spacer } from '@/components/atoms';
+import { Input, Spacer } from '@/components/atoms';
 import { getCurrencySymbol } from '@/utils/common/helper_functions';
 import { PriceTier } from './UsagePricingForm';
+import { AddChargesButton } from './SetupChargesSection';
 
 interface Props {
 	tieredPrices: PriceTier[];
@@ -148,9 +149,7 @@ const VolumeTieredPricingForm: FC<Props> = ({ setTieredPrices, tieredPrices, cur
 				</table>
 			</div>
 			<div className='flex justify-between items-center mt-4'>
-				<Button onClick={addTieredPrice} variant='default' className='flex items-center gap-2'>
-					Add Tier
-				</Button>
+				<AddChargesButton onClick={addTieredPrice} label='Add Tier' />
 			</div>
 		</div>
 	);
